feat(import): bulk import projects

+++ src/livecodes/app.ts
@@ -11,7 +11,7 @@ import {
   getLanguageByAlias,
   mapLanguage,
 } from './languages';
-import { createStorage } from './storage';
+import { createStorage, Item } from './storage';
 import {
   API,
   Cache,
@@ -59,6 +59,7 @@ import {
   compress,
   copyToClipboard,
   debounce,
+  fetchWithHandler,
   getDate,
   stringify,
   stringToValidJson,
@@ -814,7 +815,7 @@ export const app = async (appConfig: Readonly<Config>, baseUrl: string): Promise
     }
   };
 
-  const attachEventListeners = (editors: Editors) => {
+  const attachEventListeners = () => {
     const handleTitleEdit = () => {
       const projectTitle = UI.getProjectTitleElement();
       if (!projectTitle) return;
@@ -1258,9 +1259,18 @@ export const app = async (appConfig: Readonly<Config>, baseUrl: string): Promise
         const list = document.createElement('ul') as HTMLElement;
         list.classList.add('open-list');
 
-        // const bulkImportButton = UI.getBulkImportButton(listContainer);
+        const bulkImportButton = UI.getBulkImportButton(listContainer);
         const exportAllButton = UI.getExportAllButton(listContainer);
         const deleteAllButton = UI.getDeleteAllButton(listContainer);
+
+        eventsManager.addEventListener(
+          bulkImportButton,
+          'click',
+          () => {
+            showScreen('import');
+          },
+          false,
+        );
 
         eventsManager.addEventListener(
           exportAllButton,
@@ -1282,12 +1292,14 @@ export const app = async (appConfig: Readonly<Config>, baseUrl: string): Promise
           deleteAllButton,
           'click',
           () => {
-            storage.clear();
-            penId = '';
-            if (list) list.remove();
-            if (noDataMessage) listContainer.appendChild(noDataMessage);
-            deleteAllButton.classList.add('hidden');
-            exportAllButton.classList.add('hidden');
+            notifications.confirm('Delete all saved projects?', () => {
+              storage.clear();
+              penId = '';
+              if (list) list.remove();
+              if (noDataMessage) listContainer.appendChild(noDataMessage);
+              deleteAllButton.classList.add('hidden');
+              exportAllButton.classList.add('hidden');
+            });
           },
           false,
         );
@@ -1370,9 +1382,11 @@ export const app = async (appConfig: Readonly<Config>, baseUrl: string): Promise
         const importButton = UI.getUrlImportButton(importContainer);
         eventsManager.addEventListener(importForm, 'submit', async (e) => {
           e.preventDefault();
+          const buttonText = importButton.innerHTML;
           importButton.innerHTML = 'Loading...';
           importButton.disabled = true;
-          const url = UI.getUrlImportInput(importContainer).value;
+          const importInput = UI.getUrlImportInput(importContainer);
+          const url = importInput.value;
           const imported = await importCode(url, {}, defaultConfig);
           if (imported && Object.keys(imported).length > 0) {
             await loadConfig(
@@ -1382,73 +1396,121 @@ export const app = async (appConfig: Readonly<Config>, baseUrl: string): Promise
               },
               location.origin + location.pathname + '#' + url,
             );
+            modal.close();
           } else {
+            importButton.innerHTML = buttonText;
+            importButton.disabled = false;
             notifications.error('failed to load URL');
+            importInput.focus();
           }
-          modal.close();
         });
 
         const importJsonUrlForm = UI.getImportJsonUrlForm(importContainer);
         const importJsonUrlButton = UI.getImportJsonUrlButton(importContainer);
         eventsManager.addEventListener(importJsonUrlForm, 'submit', async (e) => {
           e.preventDefault();
+          const buttonText = importJsonUrlButton.innerHTML;
           importJsonUrlButton.innerHTML = 'Loading...';
           importJsonUrlButton.disabled = true;
-          const url = UI.getImportJsonUrlInput(importContainer).value;
-          const fileConfig = await fetch(url)
+          const importInput = UI.getImportJsonUrlInput(importContainer);
+          const url = importInput.value;
+          fetchWithHandler(url)
             .then((res) => res.json())
+            .then((fileConfig) =>
+              loadConfig(fileConfig, location.origin + location.pathname + '?config=' + url),
+            )
+            .then(() => modal.close())
             .catch(() => {
-              modal.close();
-              notifications.error('failed to load URL');
-              return;
+              importJsonUrlButton.innerHTML = buttonText;
+              importJsonUrlButton.disabled = false;
+              notifications.error('Error: failed to load URL');
+              importInput.focus();
             });
-          if (fileConfig) {
-            await loadConfig(fileConfig, location.origin + location.pathname + '?config=' + url);
-          }
-          modal.close();
         });
 
-        const fileInput = UI.getImportFileInput(importContainer);
+        const bulkImportJsonUrlForm = UI.getBulkImportJsonUrlForm(importContainer);
+        const bulkimportJsonUrlButton = UI.getBulkImportJsonUrlButton(importContainer);
+        eventsManager.addEventListener(bulkImportJsonUrlForm, 'submit', async (e) => {
+          e.preventDefault();
+          const buttonText = bulkimportJsonUrlButton.innerHTML;
+          bulkimportJsonUrlButton.innerHTML = 'Loading...';
+          bulkimportJsonUrlButton.disabled = true;
+          const importInput = UI.getBulkImportJsonUrlInput(importContainer);
+          const url = importInput.value;
+          fetchWithHandler(url)
+            .then((res) => res.json())
+            .then(insertItems)
+            .catch(() => {
+              bulkimportJsonUrlButton.innerHTML = buttonText;
+              bulkimportJsonUrlButton.disabled = false;
+              notifications.error('Error: failed to load URL');
+              importInput.focus();
+            });
+        });
 
-        eventsManager.addEventListener(fileInput, 'change', () => {
-          if (fileInput.files?.length === 0) return;
+        const loadFile = <T>(input: HTMLInputElement) =>
+          new Promise<T>((resolve, reject) => {
+            if (input.files?.length === 0) return;
 
-          const file = (fileInput.files as FileList)[0];
+            const file = (input.files as FileList)[0];
 
-          const allowedTypes = ['application/json', 'text/plain'];
-          if (allowedTypes.indexOf(file.type) === -1) {
-            modal.close();
-            notifications.error('Error : Incorrect file type');
-            return;
-          }
-
-          // Max 2 MB allowed
-          const maxSizeAllowed = 2 * 1024 * 1024;
-          if (file.size > maxSizeAllowed) {
-            modal.close();
-            notifications.error('Error : Exceeded size 2MB');
-            return;
-          }
-
-          const reader = new FileReader();
-
-          eventsManager.addEventListener(reader, 'load', async (event: any) => {
-            const text = (event.target?.result as string) || '';
-            try {
-              await loadConfig(JSON.parse(text));
-            } catch (error) {
-              notifications.error('Invalid configuration file');
+            const allowedTypes = ['application/json', 'text/plain'];
+            if (allowedTypes.indexOf(file.type) === -1) {
+              reject('Error: Incorrect file type');
+              return;
             }
 
-            modal.close();
+            // Max 2 MB allowed
+            const maxSizeAllowed = 2 * 1024 * 1024;
+            if (file.size > maxSizeAllowed) {
+              reject('Error: Exceeded size 2MB');
+              return;
+            }
+
+            const reader = new FileReader();
+            eventsManager.addEventListener(reader, 'load', async (event: any) => {
+              const text = (event.target?.result as string) || '';
+              try {
+                resolve(JSON.parse(text));
+              } catch (error) {
+                reject('Invalid configuration file');
+              }
+            });
+
+            eventsManager.addEventListener(reader, 'error', () => {
+              reject('Error: Failed to read file');
+            });
+
+            reader.readAsText(file);
           });
 
-          eventsManager.addEventListener(reader, 'error', () => {
-            modal.close();
-            notifications.error('Error : Failed to read file');
-          });
+        const insertItems = (items: Item[]) => {
+          if (Array.isArray(items) && items.every((item) => item.pen)) {
+            storage.bulkInsert(items);
+            notifications.success('Import Successful!');
+            showScreen('open');
+            return;
+          }
+          return Promise.reject('Error: Invalid file');
+        };
 
-          reader.readAsText(file);
+        const fileInput = UI.getImportFileInput(importContainer);
+        eventsManager.addEventListener(fileInput, 'change', () => {
+          loadFile<Config>(fileInput)
+            .then(loadConfig)
+            .then(modal.close)
+            .catch((message) => {
+              notifications.error(message);
+            });
+        });
+
+        const bulkFileInput = UI.getBulkImportFileInput(importContainer);
+        eventsManager.addEventListener(bulkFileInput, 'change', () => {
+          loadFile<Item[]>(bulkFileInput)
+            .then(insertItems)
+            .catch((message) => {
+              notifications.error(message);
+            });
         });
 
         modal.show(importContainer, { isAsync: true });
@@ -1972,7 +2034,7 @@ export const app = async (appConfig: Readonly<Config>, baseUrl: string): Promise
         },
       ];
       toolsPane = createToolsPane(toolList, getConfig(), baseUrl, editors, eventsManager);
-      attachEventListeners(editors);
+      attachEventListeners();
     } else {
       await updateEditors(editors, getConfig());
     }
